# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: ModuForge API
  description: Backend API for ModuForge — Magisk/KernelSU/APatch module development platform
  version: "2.0.0"
  contact:
    name: ModuForge Team
    url: https://moduforge.app
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
...
security:
  - BearerAuth: []
...
paths:
...
  /projects/{id}/build:
    post:
      tags: [Build]
      summary: 创建构建
      security: [{ BearerAuth: [] }]
...
  /builds/{id}:
    get:
      tags: [Build]
      summary: 获取构建详情
      security: [{ BearerAuth: [] }]
...
  /projects:
    get:
      tags: [Projects]
      summary: 项目列表
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: 项目列表
    post:
      tags: [Projects]
      summary: 创建项目
      security: [{ BearerAuth: [] }]
...